Restructure Cydia for browsing by Sources.
Using the new filtering queue, restructure Cydia to use the concept of the Source, rather than the Section as the top level concept. This removes SourceController and modifies SectionController and SectionsController to accept a Source, in addition to their existing filtering parameters (this Source can be set to nil to show all sources, although that functionality is currently untested).

In addition, the cydia:// URL schemes are modified in the following way:
 - cydia://sections works as before, showing all sections from all sources.
 - cydia://sources/<source> now shows the list of sections for a source.
 - cydia://sources/<source>/<section> has been added to show a particular section inside a source.

 This patch does not include a way in the UI to view packages from all sources. If this is desired, it can likely be very easily implemented by simply adding a row to the sources list that sets the Source to nil on the SectionsController when tapped.